requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
rich==13.7.0
click==8.1.7
requests==2.31.0
beautifulsoup4==4.12.2
ollama==0.6.1

# Optional external AI APIs
# Uncomment if you want to use external AI services
# openai==1.3.7
# anthropic==0.7.7
# google-generativeai==0.3.2

# Optional for future voice features
# whisper==1.1.10
# pyttsx3==2.90
